DSND-2047: Fix findAppointments sorting issues
Reverted to the original repository.findOfficerAppointments() implementation for the
general case, but any resource limit exceeded exceptions thrown will the use the new
two-stage findOfficerAppointmentsSparseAggregate()/findOfficerAppointmentsInIdList()
methods.

These two approaches may return results in a different order, where elements that
are equal in terms of their sort ordering, but are in different locations in the result
set. This difference may be due to the indexes used to do the initial fetch and match.